Refactor uml connectors to a UmlConnectorView
